Rework partition table layout for Limine 6
Limine 6 drops support for Ext2/3/4 and requires that boot files (incl
config, limine-bios.sys, the kernel, etc.) be stored on for example a
FAT32 partition.

As such, we make a new FAT32 /boot partition, which for EFI images
doubles as the ESP, and exists for GRUB BIOS/MBR layouts for consistency.